Cronos V6.0 Build script
* Massive overhaul to the build script, to remove redundant code and reduce clutter
* Add BUILD_COMPILER function to switch between GCC and CLANG (For future)
* Add BUILD_VAR to organize target features (AOSP, OneUI Q , Treble etc) and reduce clutter
* Add BUILD function to simplify device-specific variables and not to mix them with target variables
* Add BUILD_ALL function to simplify bulding all 4 variants with the same arguments without redundant steps
* Redo and simplify the build menu
* add comments on most steps to explain what everything does
* Clean dirty build status
* rename variants to their codename and not model name

This will be my script moving forward for all of my active kernels. it does not have to be as complicated as before. and over all it is much easier to read and expand upon.
you only need to add your device-variables into BUILD func and everything else / ramdisk and so on will be taken care of

TODO : Add BUILD_FLASHABLE function & put device variables outside of main BUILD command. would be better to keep the build command for calling other functions only

Signed-off-by: Anan Jaser <ananjaser@gmail.com>